requirements.txt
django==6.0.3
djangorestframework==3.17.1
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.9.0
psycopg2-binary==2.9.11
pillow==12.1.1
dj-database-url==3.0.1
gunicorn==23.0.0
whitenoise==6.11.0
